[JsConsoleView] made line-breaking a lot saner
Seems I had tried to solve the same problem with different methods in two
different places. The result was an unholy mess. Removed one of the solutions,
making the result readable again. Was the solution I removed used somewhere?
Hope not. If you track your problems down to this commit, it probably was...

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@9697 fcb5ba71-d80d-0410-8237-ba3920747fcc