moved topLevelCommands JsConsoleView -> JsThread
The topLevelCommands {clear, print, say} are now initialized when the
JsThread (re)boots and gets itself a fresh JsEnvironment. Since JsThread
doesn't know about JsConsoleView, the data structure was moved to the
former. This way, the commands are not lost after a forced cancel/reboot.

It is possible that topLevelCommands should really reside in JsEnvironment
instead. Time will tell.

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@10068 fcb5ba71-d80d-0410-8237-ba3920747fcc