[n.b.doc] nwo-tolerant class/namespace finding
Added the ability for the doc generator to assess the relationship
between a manager interface and its implementing classes and their
namespaces, through the xml files in the META-INF/spring directories
which contain this information for 'New World Order'-type managers.

The old mechanism is also kept, as there are managers which still
use it.

The UI manager is the only manager for which class/namespace is not
found after applying this patch. Will look into that.

git-svn-id: https://bioclipse.svn.sourceforge.net/svnroot/bioclipse/bioclipse2/trunk@11642 fcb5ba71-d80d-0410-8237-ba3920747fcc