Fix for bug 1798
Adds method getAnnotationById so the annotation can easily be reloaded after structuredb.annotate, which just changes things in the database, has been called. One strange thing is that a size() call seems to be needed for things to work. Without it the annotation looses all it's molecules during a save call. No idea why this is.